+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Update Gradle to latest stable 7.2
 - [android] Update Android Gradle plugin and Android target version
 
+## [1.0.1] - 2021-12-20
+
 ## [1.0.0] - 2021-07-02
 ### Added
 - Add methods that generate an Astarte Device ID using either a random UUIDv4 or by deriving an UUIDv5 using unique data.